refactor(backends): extract _codex_fs_atomic.py from _codex_session_storage.py
Move stateless filesystem primitives (lexists/fsync/atomic_json/write_reconciliation_audit/
read_bounded/ensure_directory_chain/decode_mount_path/filesystem_mount_root/
filesystem_type/replace_symlink) to _codex_fs_atomic.py. The transaction-
boundary core (CodexSessionStore, CodexInteractiveSessionLease, _FileLease)
remains in _codex_session_storage.py.

Update DURABLE_ARTIFACT_WRITERS registry and AST-target guard in
test_durable_artifact_writers_guard.py to point at the new module
location. Narrow E13 exemption to 1500 lines to accommodate the
remaining core after extraction.

Refs: #4664
